// Loadshop.Web.UI/src/app/shared/models/shipping-load-view.ts
import { UserContactData } from '.';

export interface ShippingLoadView {
  loadId: string;
  customerId: string;
  userLaneId: string;
  referenceLoadId: string;
  referenceLoadDisplay: string;
  originLat: number;
  originLng: number;
  originCity: string;
  originState: string;
  originEarlyDtTm: Date;
  originLateDtTm: Date;
  destLat: number;
  destLng: number;
  destCity: string;
  destState: string;
  destEarlyDtTm: Date;
  destLateDtTm: Date;
  equipmentId: string;
  equipmentType: string;
  equipmentCategoryId: string;
  equipmentCategoryDesc: string;
  equipmentTypeDisplay: string;
  stops: number;
  miles: number;
  smartSpotRate: number;
  lineHaulRate: number;
  shippersFSC: number;
  commodity: string;
  weight: number;
  isHazMat: boolean;
  transactionTypeId: string;
  distanceFromOrig: number;
  distanceFromDest: number;
  scac: string;
  bookedUser: UserContactData;
  bookedUserCarrierName: string;
  billingLoadId: string;
  billingLoadDisplay: string;
  visibilityPhoneNumber: string;
  visibilityTruckNumber: string;
  mobileExternallyEntered: boolean;
  visibilityChgDtTm: Date;
  showVisibilityWarning: boolean;
  isEstimatedFSC: boolean;
  customerLoadTypeId: number;
  serviceTypeIds: number[];
}

export const defaultShippingLoadView: ShippingLoadView = {
  loadId: null,
  customerId: null,
  userLaneId: null,
  referenceLoadId: null,
  referenceLoadDisplay: null,
  originLat: null,
  originLng: null,
  originCity: null,
  originState: null,
  originEarlyDtTm: null,
  originLateDtTm: null,
  destLat: null,
  destLng: null,
  destCity: null,
  destState: null,
  destEarlyDtTm: null,
  destLateDtTm: null,
  equipmentId: null,
  equipmentType: null,
  equipmentCategoryId: null,
  equipmentCategoryDesc: null,
  equipmentTypeDisplay: null,
  stops: null,
  miles: null,
  smartSpotRate: null,
  lineHaulRate: null,
  shippersFSC: null,
  commodity: null,
  weight: null,
  isHazMat: null,
  transactionTypeId: null,
  distanceFromOrig: null,
  distanceFromDest: null,
  scac: null,
  bookedUser: null,
  bookedUserCarrierName: null,
  billingLoadId: null,
  billingLoadDisplay: null,
  visibilityPhoneNumber: null,
  visibilityTruckNumber: null,
  mobileExternallyEntered: null,
  visibilityChgDtTm: null,
  showVisibilityWarning: null,
  isEstimatedFSC: null,
  customerLoadTypeId: null,
  serviceTypeIds: null,
};